requirements.txt
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Dance==6.2.0
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
marshmallow==3.18.0
marshmallow-sqlalchemy==0.28.1
oauthlib==3.2.2
packaging==21.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==1.4.43
urllib3==1.26.12
URLObject==2.4.3
Werkzeug==2.2.2